Added a fromRef<Message> function to be able to convert a ref-message (view) to a standard message
This is useful to store messages in a container. For example, to be able to store messages
in a std::vector<boost::any> containers, with Ref-message types, special care must be taken
to also store the same buffer that was used to visit the frame, otherwise bad things will happen
when attempting to get() or tryGet() a tag.

Thus, instead of storing Ref-message types, it is now possible to first convert a Ref-message
into its owning version (non-ref), prior to storing it in a container.