Fixed Checksum and BodyLength calculation.
In order to avoid traversing the whole frame to compute the Checksum, we were relying
on the xsputn overload of std::basic_streambuf to acculate the sum of all ascii
characters while we were writing the frame. That way, we could compute everything
(BodyLength and Checksum) in a single pass.

However, xsputn is only called when writing strings into the streambuf. When writing
fields like ints or doubles, xputc is called, which can not be overriden. This means
that we were only accumulating string characters to compute checksum and size.

To compute the checksum, we are now re-traversing the whole frame after serializing it,
that way we can accumulate all the characters, including tags and non-string values.

While this is not ideal, serialization latency is not a big concern as of today. As
a workaround of xputc not being overridable, overflow could potentially be used as
a replacement mechanism. However, this implies manually handling the pbase, pptr
end epptr pointers.

A unit-test has been added to validate that the BodyLength and Checksum are now
correctly calculated